added JS functionality, modularization, and light styling
