+++ src/index.html
@@ -7,10 +7,17 @@
 </head>
 <body>
     <!-- header -->
-    <h1>BEARS</h1>
+    <div>
+      <div class="headerText">
+        <h1>Brown Bears</h1>
+        <h4>of Katmai National Park</h4>
+        <button type="button" class="btn btn-outline-light" id="showForm">Build a Bear</button>
+      </div>
+    <img src='./images/header-pic.jpg' class='bearHeader' alt='bearycool'>
+    </div>
     
     <!-- form -->
-    <div class="card card-body w-25 p-3">
+    <div class="card card-body" id="build-a-bear">
       <h3>Build a Bear</h3>
       <form id="bear-form">
         <div class="mb-3">
@@ -19,9 +26,9 @@
         </div>
         <div class="mb-3">
           <label for="imageUrl" class="form-label">Image Url</label>
-          <input type="url" class="form-control" id="imageUrl" required>
+          <input type="url" class="form-control" id="imgUrl" required>
         </div>
-        <button type="submit" class="btn btn-primary">Submit</button>
+        <button type="submit" class="btn btn-primary" id="submit">Submit</button>
   </div>
 </div>
     <!-- card -->
