+++ src/javascripts/helpers/buttonEvents.js
@@ -1,0 +1,7 @@
+import getBearStuff from './getBearStuff';
+
+const buttonEvents = () => {
+  document.querySelector('#submit').addEventListener('click', getBearStuff);
+};
+
+export default buttonEvents;
