+++ src/javascripts/helpers/data/bearsData.js
@@ -1,0 +1,3 @@
+const bears = [];
+
+export default bears;
