+++ src/styles/main.scss
@@ -1,9 +1,15 @@
 @import "~bootstrap/scss/bootstrap";
 @import "~@fortawesome/fontawesome-free/css/all.min.css";
 
-body {
-  background-color: antiquewhite;
-  color: brown;
-  text-align: center;
-  margin-top: 100px;
+.bearHeader {
+  width: 100%;
+  height: 300px;
+  object-fit:cover;
+}
+
+.headerText {
+  position: absolute;
+  top: 80px;
+  left: 30px;
+  color: white;
 }
